点击读完啦功能实现，和插入书评返回书评的id号

+++ src/main/java/com/mark/backend/service/IRemarkService.java
@@ -20,5 +20,6 @@ public interface IRemarkService {
 	 * @param remark
 	 * @return
 	 */
-	public Integer createRemark(Remark remark);
+	public Long createRemark(Remark remark);
+
 }
